pyproject.toml
[tool.poetry]
name = "course-work-5"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^25.1"
redis = "^5.2.1"
celery = "^5.4.0"
eventlet = "^0.39.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.8"
coverage = "^7.6.12"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
